requirements.txt
Flask==3.0.3
gunicorn==23.0.0
Werkzeug==3.0.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
pandas==2.1.4
